Error with single letters fixed
When a letter was not a word it caused problems with expectedpassword in the end. Fixed with checking if expectedpassword length is 1 and same letter as in savewords. Then it will just repalce the letter. So now the for loop is not cut of early when it does not know what to do with single letters caught in savewords